ForwardPipeline/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=0.12.0=pypi_0
blas=1.0=mkl
ca-certificates=2021.4.13=hecd8cb5_1
cachetools=4.2.2=pypi_0
certifi=2020.12.5=py38hecd8cb5_0
chardet=4.0.0=pypi_0
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
dill=0.3.3=pypi_0
filelock=3.0.12=pypi_0
freetype=2.10.4=ha233b18_0
future=0.18.2=pypi_0
gdown=3.12.2=pypi_0
google-auth=1.30.1=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
grpcio=1.38.0=pypi_0
idna=2.10=pypi_0
imageio=2.9.0=pypi_0
imgaug=0.4.0=pypi_0
intel-openmp=2021.2.0=hecd8cb5_564
joblib=1.0.1=pypi_0
jpeg=9b=he5867d9_2
kiwisolver=1.3.1=pypi_0
lcms2=2.12=hf1fd2bf_0
libcxx=10.0.0=1
libffi=3.3=hb1e8313_2
libpng=1.6.37=ha441bb4_0
libtiff=4.1.0=hcb84e12_1
lxml=4.5.0=pypi_0
lz4-c=1.9.3=h23ab428_0
markdown=3.3.4=pypi_0
matplotlib=3.1.3=pypi_0
mkl=2021.2.0=hecd8cb5_269
mkl-service=2.3.0=py38h9ed2024_1
mkl_fft=1.3.0=py38h4a7008c_2
mkl_random=1.2.1=py38hb2f4e1b_2
multiprocess=0.70.11.1=pypi_0
ncurses=6.2=h0a44026_1
networkx=2.5.1=pypi_0
ninja=1.10.2=hf7b0b51_1
numpy=1.20.2=py38h4b4dc7a_0
numpy-base=1.20.2=py38he0bd621_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py_0
opencv-python=4.2.0.32=pypi_0
openssl=1.1.1k=h9ed2024_0
p-tqdm=1.3.3=pypi_0
pathos=0.2.7=pypi_0
pillow=8.2.0=py38h5270095_0
pip=21.1.1=py38hecd8cb5_0
pox=0.2.9=pypi_0
ppft=1.6.6.3=pypi_0
protobuf=3.17.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=2.4.7=pypi_0
pysocks=1.7.1=pypi_0
python=3.8.10=h88f2d9e_7
python-dateutil=2.8.1=pypi_0
pytorch=1.6.0=py3.8_0
pywavelets=1.1.1=pypi_0
pyyaml=5.3.1=pypi_0
readline=8.1=h9ed2024_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scikit-image=0.18.1=pypi_0
scikit-learn=0.23.2=pypi_0
scipy=1.4.1=pypi_0
setuptools=52.0.0=py38hecd8cb5_0
shapely=1.7.0=pypi_0
six=1.15.0=py38hecd8cb5_0
sqlite=3.35.4=hce871da_0
tensorboard=2.3.0=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
thop=0.0.31-2005241907=pypi_0
threadpoolctl=2.1.0=pypi_0
tifffile=2021.4.8=pypi_0
tk=8.6.10=hb0a8c7a_0
torchvision=0.7.0=py38_cpu
tqdm=4.43.0=pypi_0
ujson=1.35=pypi_0
urllib3=1.26.5=pypi_0
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
xmljson=0.2.0=pypi_0
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
